sale_order_import: fix missing company default
SO and SOL creation should be done w/ all defaults loaded.
In particular, the company is not propagated directly to the SOL
when we play onchanges.

Hence, at least in our tests, it can happen
that the company is not set yet on the line
and the computation of the price via 'product._get_tax_included_unit_price'
is broken when calling 'res.currency._convert' w/ an empty company.

Error example::

    [...]/base/models/res_currency.py", line 192, in _convert
    assert company, "convert amount from unknown company"
    AssertionError: convert amount from unknown company

ATM I'm adding only the company since it turns out that
if we load all defaults w/ `order.default_get` it breaks price
computation.